Update patch set 1
Patch Set 1:

(2 comments)

I see these are just copied from system/extras/simpleperf. It'd be nice to just copy those as part of the build, but I'd rather not add that to the manifest since that would mean a *lot* of extra builds for the NDK. Maybe they should be build artifacts downloaded with the rest of the prebuilts?

Patch-set: 1
Label: Presubmit-Verified=0